- Move wallet validation to cmd/client/wallet/main.go - Move validation test to cmd/client/wallet/validation_test.go
